+++ CHANGES.rst
@@ -9,7 +9,26 @@ Changes
   [rnix, 2017-02-21]
 
 
-1.0a10 (unreleased)
+1.0a11 (unreleased)
+-------------------
+
+- Add ``cone.light.browser.exception.not_found_view`` and ``not_found`` tile.
+  [rnix, 2018-09-11]
+
+- Move ``cone.light.browser.login.forbidden_view`` to
+  ``cone.app.browser.exception``.
+  [rnix, 2018-09-11]
+
+- Move registration of ``unauthorized`` tile from ``cone.app.browser.layout``
+  to ``cone.app.browser.exception``.
+  [rnix, 2018-09-11]
+
+- Move ``cone.app.browser.exception.format_traceback`` to
+  ``cone.app.browser.utils``.
+  [rnix, 2018-09-11]
+
+
+1.0a10 (2018-07-17)
 -------------------
 
 - Modify response body of ``request.response`` and return this one instead of